[MU4] Fix compiler warnings

+++ src/framework/shortcuts/internal/shortcutsregister.h
@@ -96,7 +96,7 @@ private:
 namespace std {
 inline uint qHash(const std::string& key, uint seed = 0)
 {
-    return ::qHash(QByteArray::fromRawData(key.data(), key.length()), seed);
+    return ::qHash(QByteArray::fromRawData(key.data(), static_cast<int>(key.length())), seed);
 }
 }
